#259 reorganize tests for clarity
